net: bridge: locally receive all multicast packets if IFF_ALLMULTI is set
If multicast snooping is enabled, multicast packets may not always end up
on the local bridge interface, if the host is not a member of the multicast
group. Similar to how IFF_PROMISC allows all packets to be received
locally, let IFF_ALLMULTI allow all multicast packets to be received.

Signed-off-by: Felix Fietkau <nbd@nbd.name>
Signed-off-by: Shengyu Qu <wiagn233@outlook.com>
---
Since Felix didn't send v2 for this patch, I decided to do it by myself.

Changes since v1:
 - Move to net-next
 - Changed code according to Nikolay's advice